Move convenience functions to _rtools.
